Fix query imports
