+++ database/notes.ts
@@ -24,3 +24,37 @@ export const createNote = cache(
     return note;
   },
 );
+
+export const getNotesBySessionToken = cache(async (token: string) => {
+  const notes = await sql<Note[]>`
+    SELECT
+      notes.*
+    FROM
+      notes
+      INNER JOIN sessions ON (
+        sessions.token = ${token}
+        AND notes.user_id = sessions.user_id
+        AND sessions.expiry_timestamp > now()
+      )
+  `;
+  return notes;
+});
+
+export const getNoteBySessionToken = cache(
+  async (token: string, noteId: number) => {
+    const [note] = await sql<Note[]>`
+      SELECT
+        notes.*
+      FROM
+        notes
+        INNER JOIN sessions ON (
+          sessions.token = ${token}
+          AND notes.user_id = sessions.user_id
+          AND sessions.expiry_timestamp > now()
+        )
+      WHERE
+        notes.id = ${noteId}
+    `;
+    return note;
+  },
+);
