+++ database/users.ts
@@ -1,7 +1,6 @@
 import { cache } from 'react';
 import { sql } from '../database/connect';
 import { User } from '../migrations/00006-createTableUsers';
-import { Note } from '../migrations/00008-createTableNotes';
 
 export type UserWithPasswordHash = User & {
   passwordHash: string;
@@ -67,37 +66,3 @@ export const getUserBySessionToken = cache(async (token: string) => {
   `;
   return user;
 });
-
-export const getNotesBySessionToken = cache(async (token: string) => {
-  const notes = await sql<Note[]>`
-    SELECT
-      notes.*
-    FROM
-      notes
-      INNER JOIN sessions ON (
-        sessions.token = ${token}
-        AND notes.user_id = sessions.user_id
-        AND sessions.expiry_timestamp > now()
-      )
-  `;
-  return notes;
-});
-
-export const getNoteBySessionToken = cache(
-  async (token: string, noteId: number) => {
-    const [note] = await sql<Note[]>`
-      SELECT
-        notes.*
-      FROM
-        notes
-        INNER JOIN sessions ON (
-          sessions.token = ${token}
-          AND notes.user_id = sessions.user_id
-          AND sessions.expiry_timestamp > now()
-        )
-      WHERE
-        notes.id = ${noteId}
-    `;
-    return note;
-  },
-);
